Moved PLuginsArea to separate file
