+++ edit-post/api/index.js
@@ -1,4 +1,7 @@
-import { PluginRegistry } from './plugin';
+/**
+ * Internal dependencies
+ */
+import PluginRegistry from './plugin';
 import { PluginSidebarFill as PluginSidebar } from '../components/plugin-sidebar';
 
 const registry = PluginRegistry.getInstance();
